Changing Doxyfile so that doxygen will scan files matching the pattern *.dox.
Also adding files src/README.dox indicating how to add documentation
to the "main page", and src/mpi/README.dox as an example of adding
a page of supplementary information.

Don't know if there is a "standard" way of doing this, but it seems
to work, and saves us having to stuff too much detail into real
source files.