A change to the ompi_list system that *should* be transparent. It shows no affect on anything in the unit test, and works with everything I have used to test it.
The change is to the destructor for ompi_list_t - it now cleans releases all objects attached to the list. Before this change, you had to be careful to run the list of items yourself, releasing all of them, before calling OBJ_DESTRUCT on the head of the list - or else "leak" all the item memory. With this change, you don't have to do that yourself. The destructor takes care of it for you.

Please let me know if anyone observes undesirable behavior from the change.
Ralph