Since everyone has to recompile almost everything because of changes to the docs in src/utl/output.h from the last commit, might as well do this one.  No real rush to update to this, but you'll need to re-autogen.sh, etc.
Fixes an old, small bug that caused the static-components.h files to
always be regenerated, even if they didn't change.  This caused
several files to be unnecessarily recompiled if you ran the following:

./configure
make
./configure
make

Granted, this is a small error, but a) we have it Right in other
places, b) AC does the Right things for AC_CONFIG_HEADER, so we should
do the Right things too.