Fix a confusing situation, and eliminate a potential memory leak. MPI-1 does not say whether an errorhandler obtained via MPI_ERRHANDLER_GET (or MPI_COMM_GET_ERRHANDLER or ...) needs to be ERRHANDLER_FREE'd or not.  I don't think it does, but some users will probably disagree.  So per the lengthy comment in src/mpi/c/comm_get_errhandler.c, we increment the refcount whenever GET is invoked.
This can lead to a memory leak in FINALIZE if the user GET's some
intrinsic error handlers -- we'll increment the refcount, which means
that it will never get to 0 during FINALIZE like it should.  So add a
little extra logic to ensure that they are OMPI_OBJ_DESTROY'ed during
FINALIZE like they should be.