Update the context parser to include a callback function for options requiring additional help. Include the parser in orterun. Add Jeff's new options to both the orterun and mpirun context table.
Jeff - I wasn't sure of the operation of the "x" command line option. It looks like there are multiple parameters for it? Right now, the parser doesn't support that - obviously, it needs to do so. Seems to me that we need a flag in the context table to indicate multiple parameters are possible for a given option, then install a for loop to find them and install them in an array at the specified location. Correct? If so, I can implement - just want to check my understanding with you.

I can't test orterun here, so someone will need to ensure it still works properly.