Add two options to build_tarball.pl, which should ease building: --no_check: do not make check -- may not be interesting,   we still want to do further testing. --install-dir: On clusters, tmp may not be shared,   still, one may not want to use --scrdir
build_tests.pl is taken form build_tarball.pl.
Actually all subs are the same (is it possible to have
a module that is included by both?)
It configures and compiles two tests (now PMB and the
mpichtests) and installs them for later execution. 
This takes a (misnamed) input --outfile, the output
file of build_tarball.pl, so the test-files are
compiled for every combination of OpenMPI-compilations.

The following ompi_crontab.sh script the pulls all
combinations of MPI-tests compiled and qsubs them.