Original patch submitted by: Ralph Castain Original patch reviewed by: Jeff Squyres Revised patch submitted by: Jeff Squyres Revised patch reviewed by: Ralph Castain
Original:
Short explanation: fix how we handle the "localhost" entry in the
hostfile so that the Mac (and other multi-NIC systems) will work.

Long explanation: Jeff and I spent some time chasing this down today
(mostly Jeff), and found that the Mac was having problems with the
replacement of "localhost" with the local nodename when we read the
hostfile. Jeff then found that the Linux documentation specifically
warns about the vaguery of the value returned for "nodename" (see the
man page for uname for details). Sooo....when we replaced "localhost"
with the local "nodename", the system couldn't figure out what node we
were referring to when we tried to launch.

Solution (borrowed from LAM): if the user includes "localhost" in the
hostfile, then we do NOT allow any other entries in the hostfile - the
presence of another entry will generate an error message and cause
mpirun to gracefully exit. Obviously, then, if "localhost" is
specified in the hostfile, then we are running the application
locally.

Revised:
- Properly propagate error upwards if we have a losthost+other_node
  error
- Added logic to handle multiple instances of the same hostname
- Added logic to properly increment the slot count for multiple
  instances.  For example, a hostfile with:

  foo.example.com
  foo.example.com slots=4
  foo.example.com slots=8

  would result in a single host with a slot count of 13 (i.e., if no
  slot count is specified, 1 is assumed)
- Revised the localhost logic a bit -- some cases are ok (e.g.,
  specifying localhost multiple times is ok, as long as there are no
  other hosts)
- Removed some cellid logic that is not relevant to the beta
  (currently #if 0'ed out on the trunk, but will eventually be
  enabled)