Submitted by: George Bosilca Reviewed by: Jeff Squyres
Remove a compiler warning about the type variable being use
uninitialized.  It cannot happend but I prefer to keep the compiler
quiet.  Remove few useless lines of code (were protected by a #if 0
anyway) Make the initialization of the convertor used to compute the
size of the data being send quicker. A position of -1 means that the
stack is not created (anyway this convertor is not supposed to be used
for any communications).