Some changes that allow the mca directory to compile for Windows.
We should refrain from using the variable 'new' since it is a 
keyword in C++, and some compilers are really picky about it,
namely cl :)

Did a bit of funny wrapping in convertor.h so that the 
ompi_convertor_prepare function is only exported when
doing a windows build. Talked with George about this
change and we both agreed that the windows build is 
messy, but shouldn't corrupt the main line precausions.

Also wrapped convertor.h in the __cplusplus defines
to help with exporting the inline functions under 
windows.