Submitted by: Edgar Gabriel Reviewed by: Jeff Squyres
make sure, that the value for TYPE_ATTR is identical, independent of
the definition of OMPI_WANT_MPI2_ONE_SIDED

Reviewer note: this change was made on the trunk.  It is completely
harmless, and although the beta code is "done", seems like a good idea
to keep this tiny piece of functionality in sync between the trunk and
the branch.