Submitted by: Josh Hursey Reviewed by: Jeff Squyres
need to use get_first instead of get_begin since we don't want to
execute this loop if "nodes" is an empty list. get_first, in this loop
context, allows us to do just that, while get_begin doesn't.

This fixes a --host problem that appeared on the Linux PPC64 build.