Submitted by: George Bosilica Reviewed by: Jeff Squyres and Tim Woodall
Bringing over r7807 and 7810-7814 from the trunk:

r7807:

Construct the convertor attached to the receive request. This should
happens only on the first allocation of a request object.

r7810: 

Remove a double OMPI_DECLSPEC from the definition of one of the
predefined data-types.

r7811:

Remove the memcpy we can do it as we parse the datatypes in order to
increase their references.

r7812:

After a OBJ_DESTRUCT a convertor has to be in a usable state. Read the
comment for more informations.

r7813:

Solve thee problem with the reference count on the datatype (RT bug
1492). The problem is that the convertor (when prepared) increase the
reference count on the used datatype. This reference count will be
released only when the OBJ_DESTRUCT is called on a convertor. However,
having to call OBJ_CONSTRUCT and OBJ_DESTRUCT on each request every
time we want to use it (even when it come from the cache) is an
expensive operation. This can be avoided is the OBJ_DESTRUCT will
leave the convertor in exactly the same state as OBJ_CONSTRUCT. With
this approach we just have to call OBJ_CONSTRUCT for each convertor
once when we initially create the request.

r7814:

Dont mess around with the OBJ_DESTRUCT on the communicator. It's
quicker (and sa fer) to call directly the communicator cleanup
function (ompi_convertor_cleanup).