Submitted by: George Bosilca Reviewed by: Tim Woodall, Brian Barrett
The (I hope) final solution for the convertor problem. As all the PML
inherit the base send and receive request from the pml_base, we can
solve our problem if we construct the convertor attached to any
request in the pml_base_construct function. At the end of the life
time for each request (here life time is related to one utilisation,
without taking in account the cache) we release all information
attached to the convertors in the _FINI macro by calling the
ompi_convertor_cleanup.