Submitted by: Tim Woodall / Ralph Castain Reviewed by: Ralph Castain / Tim Woodall
r7893 (twoodall):
changed segment dictionary to hash table to improve
search time for reverse lookup

r7912 (rhc): Restore the pointer arrays to the registry
dictionaries. Revise the system so that the itag is equivalent to the
index into the pointer array. It already was, but it wasn't obvious
before (several functions relied upon it, but others "hid" the
relationship) - now, make it explicitly clear. Set things up so
lookups occur at max speed by just indexing into the dictionary array.