Submitted by: Jeff Squyres Reviewed by: Brian Barrett
Brought over r8008 from trunk.  This was initially not going to come
over, but it turned out to be necessary.  Specifically, there were two
commits for MPI_Op yesterday on the trunk:

r8008: remove MPI_REPLACE declarations and instantiations for
       MPI_REDUCE and fix string names (this patch)
r8010: enable MPI_*LONG_LONG* functions in intrinsic MPI_Op's

I brought r8010 over because it was clearly a bug.  I didn't bring
r8008 because it didn't seem to matter too much that we had a few
extra functions available and a few strings slightly incorrect.  But I
missed the fact that the REPLACE instantiations did not have the
LONG_LONG functions and therefore created a problem.  So removing all
of the MPI_REPLACE stuff for MPI_REDUCE indirectly avoids the problem,
and fixes some other minor issues along the way (e.g., the string
names). 

r8008 log message:
- Fix the MPI_Op intrinsic operation string names ("MPI_foo", not
  "MPI_OP_foo")
- Remove all the handlers for MPI_REPLACE for general reductions
  (it's only defined for MPI_ACCUMULATE, and ACCUMULATE is handled
  differently than the other reductions, so it's safe to make all the
  maps for REPLACE be empty)