Submitted: Josh Hursey Reviewed: Tim Woodall
New Revision: 8109

Modified:
   trunk/orte/mca/iof/base/iof_base_endpoint.c
Log:
Fix for orted / MPI_Abort problem reported from testers. They were seeing orteds
spining in orte_iof_base_flush() when running 
  intel_tests/src/MPI_Errhandler_fatal_c

When we close an endpoint by taking it out of the envent handler, we need to make
sure that it fits the criteria to pass through orte_iof_base_flush(), specificly
make sure we clean out the ep_frags list.
Note: This is more of a sanity check, since the endpoint should already be
      in this state at the point of closure.

Secondly in orte_iof_base_endpoint_read_handler(), if we determine that it is 
necessary to close the endpoint we have to "return" after doing so, otherwise
we add another frag to the endpoint which will cause it to hang in 
orte_iof_base_flush().

New Revision: 8121

Modified:
   trunk/orte/mca/iof/base/iof_base_endpoint.c
Log:
return fragments to free list