Submitted by: Jeff Squyres Reviewed by: Brian Barrett
Apparently r8134 was *not* brought over from the trunk to v1.0 during
the r8140 commit (even though the commit message for r8140 says that
it was).  Bonk.  :-(

I'm guessing that I mucked up the merges and somehow neglected to
include r8134 in the final commit for 8140.  The changes were already
reviewed and approved, so this commit finally bringing r8134 to the
v1.0 branch.  For real this time.