Submitted by: George Bosilca Reviewed by: Jeff Squyres
If pending_length is not zero the datatype engine believe that a
partial conversion is still in the buffer attached to the
convertor. We have to force it to zero for all new convertors. This
bug is still half fixed until I figure out a correct way to do it when
we clone a convertor ... Anyway, this future correction will not
affect open mpi.