Submitted by: Jeff Squyres Reviewed by: Brian Barrett
Convert MPI_Offset to be a typedef instead of a #define.  See the
following for the rationale:

http://www.open-mpi.org/community/lists/users/2005/11/0391.php

Note that mpi.h only receives AC_DEFINE's (not AC_SUBST's), so we
still have to AC_DEFINE to get the value in there -- so we AC_DEFINE
something that we give to a typedef.