Submitted by: Jeff Squyres Reviewed by: David Daniel
Remove an extra -o $@ that doesn't seem to be necessary, and actually
confuses the Lahey compiler (because it doesn't want to name a module
mpi_kinds.ompi_module).  We shouldn't need this because we're already
touching a sentinel file (mpi_kinds.ompi_module) for make
dependencies; we don't need the compiler-generated module to be named
that.