Submitted by: George Bosilca Reviewed by: Brian Barrett
r8741
We're thread safe again. The atomic compare-and-swap was not used in
the correct way allowing the descriptors to vanish. The PML was
thinking that they are in the btl_cache when they weren't ... It lead
to memory consumption on most environments when compiled with thread
enabled. After modification the latency went down by nearly 0.5
microseconds.

Simple way to trigger the bug: limit the number of maximum items in
the free list and run any communication intensive application (like
Netpipe).

r8762
As requested move back the alloc and return to macros instead of
static inlined functions.