Submitted by: Jeff Squyres Reviewed by: George Bosilca
r9698:
Don't bother to get the alignmnet of an f90 type -- the DDT engine
only uses the F77 alignments anyway.  Plus, it is currently unknown
how to reliably get the F90 alignments because compilers *may* reorder
things and/or give different alignments than the f77 tests.

r9707:
Finish the rest of r9698:
- Add a lengthy comment explaining why we don't get F90 data type
  alignments
- Remove config/f90_check.m4 and its entry in acinclude.m4