Submitted by: Tim Woodall Reviewed by: Sven Stork, Jeff Squyres
r9715
remove debug code - not required

r9749
Correction to resolve a problem related to partial reads. We were
making a copy of the receive buffer based on the iovec struct that may
have been updated during partial reads to reflect the current
offset. Need to make the copy using the base address of the buffer.

Thanks to Sven Stork for finding this.