Submitted by: George Bosilca Reveiwed by: Galen Shipman
r10298
Don't reset the convertor's position if it wasn't initialized
before. This can only happens for zero byte persistent requests.

r10299
The real solution. If the sendreq->req_send.req_bytes_packed is zero
then there is no data to be trasfered. And this is the condition which
lead to a non initialized convertor.