Submitted by: George Bosilca Reviewed by: Brian Barrett
r10364
Correctly compute the boundaries for the Fortran matrix style.

r10366
Do not unpack outside the legal boundaries of the data even if the specified
iov_len is larger than the amount of missing data.


r10367
Do not sum the checksum. Instead use the intermediary values in order to
correctly compute the final checksum. This is not a bug in the case where
both the sender and the receiver execute EXACTLY the same checksum
computations but is definitively a problem if not (such as the buffered case).

r10368
Change the name of one of the datatype parameters to match all the others.