Submitted by: Josh Hursey, Ralph Castain Reviewed by: Ralph Castain, Josh Hursey, Jeff Squyres
r10662
A partial fix for the hanging orted bugs (Ticket #177) 

When we force an application to terminate (via CTRL-C to mpirun) 
we send an out-of-band message to the orted to reap its children. 
the fork PLS was doing an internal waitpid but never releasing or 
updating the information and signaling the condition variable. So 
the fork PLS callback for SIGCHLD registered with the event library 
and this waitpid are in a bit of a race to 'waitpid' for the
children. 
Since the PLS callback was the only one that handled the signal
properly 
when it 'won' then things were great -- as in the normal termination
case. 
But when it 'lost' -- as in the abnormal termination case -- the orted 
never received the proper signal that its children had gone away. 

We want to preserve the internal fork PLS callback since it allows 
for a timeout while waiting for the child, which the event library 
won't do. 

This allows both to exist, and behave properly. 

This was introduced in r9068. 

The ticket is still open since the orted's hang in other situations 
still. This is a fix for one of the causes. 

r10666
A bit more cleanup for that last patch. 

 * num_children should really be an int instead of size_t 
since 'size_t' is not signed and num_children can (in rare cases) 
drop below 0, and don't want it to roll around to MAX_INT or some 
such. 
 * I figured out that this problem only happened to me because I use 
the pls_fork_reap_timeout MCA parameter and thus the only time that 
the code in pls_fork_module.c to waitpid is executed is if this is 
not set to 0 (I had it set to 1 to give my procs time to exit). I 
adjusted the loop from while{...} to do{...}while; so that it is 
executed at least once for consistency. 
 * de-register the SIGCHILD callback for the pid before we attempt 
to kill it, so that we don't leave the door open for both the 
waitpids (the one in the callback, and the one in this function) 
to race to see who can wait on the child. 
 * Move the 'thread release' to outside the for loop for a bit of an 
optimization, and always set the value to 0 since we want to 
finish after this function. 
 * Added a help message for the case when we can't send a kill() 
signal to the process. Should never happen, but all is possible 
in the wild wild west of HPC. 

r10866
Fix a bug noted by Jeff - we were no longer accurately recording in
the registry that a process had been terminated when the user
initiated the "kill" process (via cntrl-c). 

Added another system-level test function for ORTE that just spins
until terminated by a ctrl-c signal. 

Modified orterun - added a couple of newlines to the output when
abnormally terminating so the prompt always is on a new line.