Submitted by: David Daniel Reviewed by: Josh Hursey
Adding support for MPI_Comm_spawn() to use the 'host' key in an
MPI_Info object if provided.

The associated value is a comma-separated list of hosts -- which must
be in the initial allocation -- and is used to populate the
application context map.

(also added a NEWS item about this and re-indented some of the code in
comm_dyn.c to match the indenting in the rest of the file)