Submitted by: Jeff Squyres Reviewed by: Rolf Vandevaart
Refs #302.  Now thowing over the wall to the 1.2 gatekeepers...

r11356
A bunch of fixes for Fortran string issues. In general, ensure to
convert between fortran and C string representations properly. In
doing so, we properly adhere to the MPI spec stating that MPI_Info
keys and values must be whitespace-trimmed when coming in from
Fortran. Hence, this fixes bug #241.

r11367
Oops -- in MPI_INFO_GET_NTHKEY, the key argument is an '''out'''
value, not an '''in''' value. So the string needs to be converted c2f,
not f2c.