Refs #357
 * Submitted by: Josh Hursey
 * Reviewed by: Jeff Squyres

Fix a code clarity issue in the POE PLS. 

Allow the POE RAS to be compled for linux as well as AIX.  The POE RAS
is really a Loadleveler RAS, and IU now has a cluster that uses
Loadleveler in a Linux environment (BigRed).

This seems to be the only thing we need to do so far to run Open MPI
on BigRed. Yay :)